+++ objects/nodesStatistics.go
@@ -2,7 +2,7 @@ package objects
 
 // NodesStatistics is the structure for overall nodes statistics
 type NodesStatistics struct {
-	EmtpyNodesCount          int64 `json:"emtpyNodesCount"`
-	SynchronizedNodesCount   int64 `json:"synchronizedNodesCount"`
-	UnsynchronizedNodesCount int64 `json:"unsynchronizedNodesCount"`
+	EmtpyNodesCount          int `json:"emtpyNodesCount"`
+	SynchronizedNodesCount   int `json:"synchronizedNodesCount"`
+	UnsynchronizedNodesCount int `json:"unsynchronizedNodesCount"`
 }
